% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch COVID-19 data}
\usage{
fetch_data(location = "Jawa Barat")
}
\arguments{
\item{location}{Which location data do you want to fetch? Currently the valid values are "Jawa Barat" and "Indonesia"}
}
\value{
a tibble with columns indicating daily incidence of "positif", "sembuh", "meninggal". If the location is Jawa Barat, two additional columns are added: "odp" and "pdp".
}
\description{
Fetch COVID-19 data
}
\examples{
(jabar <- fetch_data(location = "Jawa Barat"))
(indonesia <- fetch_data(location = "Indonesia"))
}
