% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_incidence.R
\name{inspect_incidence}
\alias{inspect_incidence}
\title{Inspect incidence}
\usage{
inspect_incidence(data, which = "positif")
}
\arguments{
\item{data}{COVID-19 data.}

\item{which}{Which category do you want to plot? Valid values are "positif", "sembuh", "meninggal". For Jawa Barat Data, two additional values are valid: "odp" and "pdp".}
}
\value{
A COVID-19 incidence object.
}
\description{
Inspect incidence
}
\examples{
# Fetch data
daily <- fetch_data("Jawa Barat")

# Inspect incidence
covid_incidence <-
  inspect_incidence(daily, "positif")
covid_incidence
}
