% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_trend.R
\name{inspect_trend}
\alias{inspect_trend}
\title{Inspect trend}
\usage{
inspect_trend(incidence, si_mean = 4.8, si_sd = 2.3)
}
\arguments{
\item{incidence}{A COVID-19 incidence object.}

\item{si_mean}{The mean of serial interval.}

\item{si_sd}{The standard devation of serial interval.}
}
\value{
A tibble of instantenous reproduction number estimates.
}
\description{
Inspect trend
}
\examples{
# Fetch data
daily <- fetch_data("Jawa Barat")

# Inspect incidence
covid_incidence <-
  inspect_incidence(daily, "positif")
show_incidence_s(covid_incidence)

# Inspect daily trend
covid_trend <-
  inspect_trend(covid_incidence)
covid_trend
}
