% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_beds.R
\name{project_beds}
\alias{project_beds}
\title{Project beds occupancy}
\usage{
project_beds(projection, los = 13, n_sim = 10)
}
\arguments{
\item{projection}{COVID-19 projection.}

\item{los}{Length of stay in hospital.}

\item{n_sim}{The number of times duration of hospitalisation is simulated for each admission.}
}
\value{
A projection of beds occupancy.
}
\description{
Project beds occupancy
}
\examples{
# Fetch data
daily <- fetch_data("Jawa Barat")

# Inspect incidence
covid_incidence <-
  inspect_incidence(daily, "positif")
show_incidence_s(covid_incidence)

# Inspect daily trend
covid_trend <-
  inspect_trend(covid_incidence)
show_trend_s(covid_trend)

# Project future incidence
covid_projection <-
  project_incidence(
    covid_incidence,
    n_days = 21
  )
covid_projection

# Project beds need
beds_projection <-
  project_beds(
    projection = covid_projection,
    los = 13,
    n_sim = 10
  )
beds_projection
}
