% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_incidence.R
\name{project_incidence}
\alias{project_incidence}
\title{Project incidence}
\usage{
project_incidence(incidence, n_days = 21, si_mean = 4.8, si_sd = 2.3)
}
\arguments{
\item{incidence}{COVID-19 incidence object.}

\item{n_days}{The number of days to run simulations for.}

\item{si_mean}{The mean of serial interval.}

\item{si_sd}{The standard devation of serial interval.}
}
\value{
A list of R estimate, daily prohected incidence, and cumulative projected incidence.
}
\description{
Project incidence
}
