% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_incidence_i.R
\name{show_incidence_i}
\alias{show_incidence_i}
\title{Show interactive incidence plot}
\usage{
show_incidence_i(incidence)
}
\arguments{
\item{incidence}{A COVID-19 incidence object.}
}
\value{
A highchart object.
}
\description{
Show interactive incidence plot
}
\examples{
# Fetch data
daily <- fetch_data("Jawa Barat")

# Inspect incidence
covid_incidence <-
  inspect_incidence(daily, "positif")
show_incidence_i(covid_incidence)
}
