% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_incidence_s.R
\name{show_incidence_s}
\alias{show_incidence_s}
\title{Show incidence plot}
\usage{
show_incidence_s(incidence)
}
\arguments{
\item{incidence}{A COVID-19 incidence object.}
}
\value{
A ggplot2 object.
}
\description{
Show incidence plot
}
