% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend_i.R
\name{show_trend_i}
\alias{show_trend_i}
\title{Show interactive trend plot}
\usage{
show_trend_i(trend)
}
\arguments{
\item{trend}{A COVID-19 trend object.}
}
\value{
A highchart object.
}
\description{
Show interactive trend plot
}
\examples{
# Fetch data
daily <- fetch_data("Jawa Barat")

# Inspect incidence
covid_incidence <-
  inspect_incidence(daily, "positif")
show_incidence_s(covid_incidence)

# Inspect daily trend
covid_trend <-
  inspect_trend(covid_incidence)
show_trend_i(covid_trend)
}
