% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_projection_s.R
\name{show_projection_s}
\alias{show_projection_s}
\title{Show static projection plot}
\usage{
show_projection_s(projection, value = "cumulative")
}
\arguments{
\item{projection}{COVID-19 projection.}

\item{value}{Whether to plot daily incidence or cumulative incidence. Valid values are "daily" and "cumulative".}
}
\value{
a ggplot2 object.
}
\description{
Show static projection plot
}
\examples{
# Fetch data
daily <- fetch_data("Jawa Barat")

# Inspect incidence
covid_incidence <-
  inspect_incidence(daily, "positif")
show_incidence_s(covid_incidence)

# Inspect daily trend
covid_trend <-
  inspect_trend(covid_incidence)
show_trend_s(covid_trend)

# Project future incidence
covid_projection <-
  project_incidence(
    covid_incidence,
    n_days = 21
  )
show_projection_s(covid_projection)
}
