% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend_s.R
\name{show_trend_s}
\alias{show_trend_s}
\title{Show static trend plot}
\usage{
show_trend_s(trend)
}
\arguments{
\item{trend}{A COVID-19 trend object.}
}
\value{
A ggplot2 object.
}
\description{
Show static trend plot
}
